// src/payload-types.ts
export interface Tenant {
  id: string;
  /**
   * This is the name of the store you are creating. (e.g. John Doe's Store)
   */
  name: string;
  /**
   * This is the subdomain for the store you are creating. (e.g. [slug].xspark-ebazaar.com)
   */
  slug: string;
  /**
   * This is the image for the store you are creating.
   */
  image?: (string | null) | Media;
  /**
   * Stripe account ID associated with your store.
   */
  stripeAccountId: string;
  /**
   * You cannot create products until you have submitted your Stripe details.
   */
  stripeDetailsSubmitted?: boolean | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: string;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}